♻️ Rename the subpackage to avoid to conflicting naming with the third party library.
